+++ token-client/src/test/java/com/sap/cloud/security/xsuaa/client/XsuaaOAuth2TokenServiceClientCredentialsTest.java
@@ -1,5 +1,9 @@
 package com.sap.cloud.security.xsuaa.client;
 
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertThat;
+import static org.hamcrest.CoreMatchers.is;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -18,91 +22,78 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static com.sap.cloud.security.xsuaa.client.OAuth2TokenServiceConstants.ACCESS_TOKEN;
-import static com.sap.cloud.security.xsuaa.client.OAuth2TokenServiceConstants.REFRESH_TOKEN;
-import static com.sap.cloud.security.xsuaa.client.OAuth2TokenServiceConstants.EXPIRES_IN;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertThat;
-import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.eq;
+import static org.mockito.Mockito.any;
 
 @RunWith(MockitoJUnitRunner.class)
-public class DefaultOAuth2TokenServiceUserTokenTest {
+public class XsuaaOAuth2TokenServiceClientCredentialsTest {
 
 	OAuth2TokenService cut;
 	ClientCredentials clientCredentials;
 	URI tokenEndpoint;
 	Map<String, String> responseMap;
-	private static final String userTokenToBeExchanged = "65a84cd45c554c6993ea26cb8f9cf3a2";
 
 	@Mock
 	RestOperations mockRestOperations;
 
 	@Before
 	public void setup() {
-		cut = new DefaultOAuth2TokenService(mockRestOperations);
+		cut = new XsuaaOAuth2TokenService(mockRestOperations);
 		clientCredentials = new ClientCredentials("clientid", "mysecretpassword");
 		tokenEndpoint = URI.create("https://subdomain.myauth.server.com/oauth/token");
 
 		responseMap = new HashMap<>();
-		responseMap.put(REFRESH_TOKEN, "2170b564228448c6aed8b1ddfdb8bf53-r");
-		responseMap.put(ACCESS_TOKEN, "4d841646fcc340f59b1b7b43df4b050d"); // opaque access token
-		responseMap.put(EXPIRES_IN, "43199");
+		responseMap.putIfAbsent(ACCESS_TOKEN, "f529.dd6e30.d454677322aaabb0");
+		responseMap.putIfAbsent(OAuth2TokenServiceConstants.EXPIRES_IN, "43199");
 	}
 
 	@Test(expected = IllegalArgumentException.class)
 	public void initialize_throwsIfRestOperationsIsNull() {
-		new DefaultOAuth2TokenService(null);
+		new XsuaaOAuth2TokenService(null);
 	}
 
 	@Test
 	public void retrieveToken_throwsOnNullValues() {
 		assertThatThrownBy(() -> {
-			cut.retrieveAccessTokenViaUserTokenGrant(null, clientCredentials, userTokenToBeExchanged, null);
+			cut.retrieveAccessTokenViaClientCredentialsGrant(null, clientCredentials, null);
 		}).isInstanceOf(IllegalArgumentException.class).hasMessageStartingWith("tokenEndpointUri");
 
 		assertThatThrownBy(() -> {
-			cut.retrieveAccessTokenViaUserTokenGrant(tokenEndpoint, null, userTokenToBeExchanged, null);
+			cut.retrieveAccessTokenViaClientCredentialsGrant(tokenEndpoint, null, null);
 		}).isInstanceOf(IllegalArgumentException.class).hasMessageStartingWith("clientCredentials");
-
-		assertThatThrownBy(() -> {
-			cut.retrieveAccessTokenViaUserTokenGrant(tokenEndpoint, clientCredentials, null, null);
-		}).isInstanceOf(IllegalArgumentException.class).hasMessageStartingWith("token");
 	}
 
 	@Test(expected = OAuth2ServiceException.class)
 	public void retrieveToken_throwsIfHttpStatusUnauthorized() {
 		Mockito.when(mockRestOperations.postForEntity(any(URI.class), any(HttpEntity.class), eq(Map.class)))
 				.thenThrow(new HttpClientErrorException(HttpStatus.UNAUTHORIZED));
-		cut.retrieveAccessTokenViaUserTokenGrant(tokenEndpoint, clientCredentials,
-				userTokenToBeExchanged, null);
+		cut.retrieveAccessTokenViaClientCredentialsGrant(tokenEndpoint, clientCredentials,
+				null);
 	}
 
 	@Test(expected = OAuth2ServiceException.class)
 	public void retrieveToken_throwsIfHttpStatusNotOk() {
 		Mockito.when(mockRestOperations.postForEntity(any(URI.class), any(HttpEntity.class), eq(Map.class)))
 				.thenThrow(new HttpClientErrorException(HttpStatus.BAD_REQUEST));
-		cut.retrieveAccessTokenViaUserTokenGrant(tokenEndpoint, clientCredentials,
-				userTokenToBeExchanged, null);
+		cut.retrieveAccessTokenViaClientCredentialsGrant(tokenEndpoint, clientCredentials,
+				null);
 	}
 
 	@Test
 	public void retrieveToken() {
 		HttpHeaders expectedHeaders = new HttpHeaders();
 		expectedHeaders.add(HttpHeaders.ACCEPT, "application/json");
-		expectedHeaders.add(HttpHeaders.AUTHORIZATION, "Bearer " + userTokenToBeExchanged);
 		HttpEntity expectedRequest = new HttpEntity(expectedHeaders);
 
 		Mockito.when(mockRestOperations
-				.postForEntity(eq(createUriWithParameters("grant_type=user_token&client_id=clientid")),
-						eq(expectedRequest),
+				.postForEntity(eq(createUriWithParameters("grant_type=client_credentials&client_secret=mysecretpassword&client_id=clientid")), eq(expectedRequest),
 						eq(Map.class)))
 				.thenReturn(new ResponseEntity<>(responseMap, HttpStatus.OK));
 
-		OAuth2AccessToken accessToken = cut.retrieveAccessTokenViaUserTokenGrant(tokenEndpoint, clientCredentials,
-				userTokenToBeExchanged, null);
-		assertThat(accessToken.getRefreshToken().get(), is(responseMap.get(REFRESH_TOKEN)));
+		OAuth2AccessToken accessToken = cut.retrieveAccessTokenViaClientCredentialsGrant(tokenEndpoint,
+				clientCredentials,
+				null);
 		assertThat(accessToken.getValue(), is(responseMap.get(ACCESS_TOKEN)));
 		assertNotNull(accessToken.getExpiredAtDate());
 	}
@@ -110,8 +101,7 @@ public class DefaultOAuth2TokenServiceUserTokenTest {
 	@Test
 	public void retrieveToken_withOptionalParamaters() {
 		Mockito.when(mockRestOperations.postForEntity(
-				eq(createUriWithParameters(
-						"add-param-1=value1&add-param-2=value2&grant_type=user_token&client_id=clientid")),
+				eq(createUriWithParameters("add-param-1=value1&add-param-2=value2&client_secret=mysecretpassword&grant_type=client_credentials&client_id=clientid")),
 				any(HttpEntity.class), eq(Map.class)))
 				.thenReturn(new ResponseEntity<>(responseMap, HttpStatus.OK));
 
@@ -119,24 +109,25 @@ public class DefaultOAuth2TokenServiceUserTokenTest {
 		additionalParameters.put("add-param-1", "value1");
 		additionalParameters.put("add-param-2", "value2");
 
-		OAuth2AccessToken accessToken = cut.retrieveAccessTokenViaUserTokenGrant(tokenEndpoint, clientCredentials,
-				userTokenToBeExchanged, additionalParameters);
-		assertThat(accessToken.getRefreshToken().get(), is(responseMap.get(REFRESH_TOKEN)));
+		OAuth2AccessToken accessToken = cut.retrieveAccessTokenViaClientCredentialsGrant(tokenEndpoint,
+				clientCredentials,
+				additionalParameters);
+		assertThat(accessToken.getValue(), is(responseMap.get(ACCESS_TOKEN)));
 	}
 
 	@Test
 	public void retrieveToken_requiredParametersCanNotBeOverwritten() {
-		Mockito.when(
-				mockRestOperations.postForEntity(eq(createUriWithParameters("grant_type=user_token&client_id=clientid")),
-						any(HttpEntity.class), eq(Map.class)))
+		Mockito.when(mockRestOperations.postForEntity(eq(createUriWithParameters("grant_type=client_credentials&client_id=clientid&client_secret=mysecretpassword")),
+				any(HttpEntity.class), eq(Map.class)))
 				.thenReturn(new ResponseEntity<>(responseMap, HttpStatus.OK));
 
 		Map<String, String> overwrittenGrantType = new HashMap<>();
 		overwrittenGrantType.put(OAuth2TokenServiceConstants.GRANT_TYPE, "overwrite-obligatory-param");
 
-		OAuth2AccessToken accessToken = cut.retrieveAccessTokenViaUserTokenGrant(tokenEndpoint, clientCredentials,
-				userTokenToBeExchanged, overwrittenGrantType);
-		assertThat(accessToken.getRefreshToken().get(), is(responseMap.get(REFRESH_TOKEN)));
+		OAuth2AccessToken accessToken = cut.retrieveAccessTokenViaClientCredentialsGrant(tokenEndpoint,
+				clientCredentials,
+				overwrittenGrantType);
+		assertThat(accessToken.getValue(), is(responseMap.get(ACCESS_TOKEN)));
 	}
 
 	private URI createUriWithParameters(String queryParameterList) {
